# docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "conTIgo API",
        "description": "Autenticación cookie ACCESS HttpOnly JWT RS256. Obtenga GET /api/v1/auth/csrf y envíe su token en X-XSRF-TOKEN para todo POST, incluso login, refresh y logout. Cookies same-origin; no Bearer header. Paginación base cero, size 1..100, sort: createdAt, code, priority, state. Errores ProblemDetail: 400 validación, 401 autenticación, 403 permiso/CSRF, 404 inexistente, 409 estado/concurrencia, 413 tamaño, 429 intentos.",
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "cookieAuth": []
        }
    ],
    "paths": {
...
        "/api/v1/tickets/{id}/attachments": {
            "get": {
                "tags": [
                    "attachment-controller"
                ],
                "operationId": "list_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/View"
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto de estado, versión o unicidad",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Falta autenticación válida",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "Archivo demasiado grande",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permiso o CSRF insuficientes",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Formato no compatible con visualización",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Demasiados intentos",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "attachment-controller"
                ],
                "operationId": "upload",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "X-XSRF-TOKEN",
                        "in": "header",
                        "description": "Token de GET /api/v1/auth/csrf, junto con su cookie XSRF-TOKEN.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "required": [
                                    "file"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/View"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto de estado, versión o unicidad",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Falta autenticación válida",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "Archivo demasiado grande",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permiso o CSRF insuficientes",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Formato no compatible con visualización",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Demasiados intentos",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/tickets/{id}/attachments/{attachmentId}": {
            "get": {
                "tags": [
                    "attachment-controller"
                ],
                "operationId": "download",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "attachmentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto de estado, versión o unicidad",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Falta autenticación válida",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "Archivo demasiado grande",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permiso o CSRF insuficientes",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Formato no compatible con visualización",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Demasiados intentos",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/tickets/{id}/attachments/{attachmentId}/view": {
            "get": {
                "tags": [
                    "attachment-controller"
                ],
                "summary": "Visualizar evidencia en el navegador",
                "description": "Requiere acceso al ticket. Admite PNG, JPEG, PDF y TXT UTF-8; devuelve inline, sin caché.",
                "operationId": "view",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "attachmentId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflicto de estado, versión o unicidad",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Datos inválidos",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Falta autenticación válida",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "413": {
                        "description": "Archivo demasiado grande",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Permiso o CSRF insuficientes",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Recurso inexistente",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "415": {
                        "description": "Formato no compatible con visualización",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Demasiados intentos",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetail"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "View": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "contentType": {
                        "type": "string"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "ProblemDetail": {
                "properties": {
                    "status": {},
                    "detail": {
                        "type": "string"
                    },
                    "instance": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    }
                }
            }
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "name": "ACCESS",
                "in": "cookie"
            },
            "refreshCookie": {
                "type": "apiKey",
                "name": "REFRESH",
                "in": "cookie"
            }
        }
    }
}